refactor: move resolve_config_value to dokkupy.utils
