+++ dokkupy/plugins/registry.py
@@ -1,10 +1,4 @@
-import os
-
-
-def resolve_config_value(value):
-    if isinstance(value, dict) and 'env' in value:
-        return os.environ[value['env']]
-    return value
+from dokkupy.utils import resolve_config_value
 
 
 class Registry(object):
