+++ tests/plugins/test_registry.py
@@ -4,19 +4,6 @@ import unittest
 from unittest.mock import MagicMock, patch
 
 from dokkupy.core import Dokku
-from dokkupy.plugins.registry import resolve_config_value
-
-
-class TestResolveConfigValue(unittest.TestCase):
-    def test_returns_plain_string(self):
-        self.assertEqual(resolve_config_value('hello'), 'hello')
-
-    def test_resolves_env_dict(self):
-        with patch.dict(os.environ, {'MY_VAR': 'secret'}):
-            self.assertEqual(
-                resolve_config_value({'env': 'MY_VAR'}),
-                'secret',
-            )
 
 
 @patch('dokkupy.core.subprocess.Popen')
